init: Protected UI, Sidebar

+++ src/features/actions/delete-ticket.ts
@@ -4,21 +4,34 @@ import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 import { setCookieByKey } from "@/app/actions/cookies";
 import { ticketsPath } from "@/app/paths";
-import { fromErrorToActionState } from "@/components/form/utils/to-action-state";
+import { fromErrorToActionState, toActionState } from "@/components/form/utils/to-action-state";
 import { prisma } from "@/lib/prisma";
+import { getAuthOrRedirect } from "../auth/queries/get-auth-or-redirect";
+import { isOwner } from "../auth/utils/is-owner";
 
 export const deleteTicket = async (id: string) => {
+  const { user } = await getAuthOrRedirect();
+
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: {
+        id,
+      },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not authorized");
+    }
     await prisma.ticket.delete({
       where: {
         id,
       },
     });
   } catch (error) {
-    return fromErrorToActionState(error)
+    return fromErrorToActionState(error);
   }
 
-  revalidatePath(ticketsPath())
-  await setCookieByKey("toast", "Ticket deleted")
-  redirect(ticketsPath())
+  revalidatePath(ticketsPath());
+  await setCookieByKey("toast", "Ticket deleted");
+  redirect(ticketsPath());
 };
